Clarify how to clone repo so pushes will work
Adding a few small tweaks to the readme after making a few mistakes going through the process of submitting a blog entry. I cloned using HTTPS instead of SSH and was therefore unable to push my branch to the shared repo, so I call that out. I also initially created a fork and opened a PR there, so I added a line earlier on in the instructions about not using a fork.